requirements.txt
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-cache2==0.2.1
flower==2.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanize==4.7.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pendulum==2.1.2
Pillow==10.0.0
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pyasn1==0.5.0
pydantic==2.0.2
pydantic-core==2.1.2
pydantic-settings==2.0.2
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.13.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
tornado==6.3.2
typing-extensions==4.7.1
tzdata==2023.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
WTForms==3.0.1
